// tests/types.ts
export type ThermalSpotsAnalytics = {
  admin: string;
  thermalSpotsData: {
    id:
      | "volcanoesCount"
      | "hotspotDaysPerYearMax"
      | "industrialAreaKm2"
      | "forestAreaKm2";
    minExpectedResult: number;
    maxExpectedResult: number;
  }[];
};